Store and use per-template per-state results on each node
merge results from node in the final_results using arch

Wrapper scripts for use with the pipeline
A common libary of various wrapper functionalities
Using iterator to wrap around functions that do the actual deploy, etc.
Ruby is cool! :)

bring the latest deploy.sh from osl-jenkins#80

If a qcow2 img is already built, exit gracefully

Removing old unrequired wrapper script

Call openstack_taster wrapper rightly

sprinkling comments and todos in the pipeline script

fix unless

Handle things using rescue block when possible

Better parsing from template

Do not accept CHEF_VERSION for building images
Let it be specified in the template -- the one single place
for such customizations

For deploy scripts, get the chef_version from the template

update stage results in a proper manner.

if the template was already in the array, don't create an entry for it.
if it was not, only then create a key for it with a blank hash as the value

this prevents overriding of the results in stages where the only value that stayed
was the last stage's result and everything else was lost.